Version 0.15.0D
Just touching this (i don't have the unity editor on the computer i wrote this on so i can't test it yet, going to test during period 9)


the code I was testing is on the computer in the lab(The jump thing), I didn't put it up because there still is problems with it. (can jump an infinite amount of times.