Version .91_E
A preliminary life system for the game is implemented. After 3 dinosaurs
pass the ice cream, the game resets: all other dinosaurs and turrets are
destroyed, coins are set back to 100, lives are set back to 3 and round
starts back to 1.

BTW New naming system. Starting Feb. 8, we restart from Version
.01_letter. Change the letter every day/update. The number increments
.01 every week